refactor: ApplicationContext 분리
